add fun floating bubbles using Paper.js
I asked my mentor for fun visual ideas, and he introduced me to Paper.js. After playing around with it (and breaking a few things at first!), I figured out how to create floating bubbles on the side of the page. They’re now using the colors of the Brazilian flag to match the recipe theme, and I think they add a really cute and playful touch to the site!